feat(intake): CorpusIntake.assert_edge + relocate EdgeStore (#263)
ADR-0006 Slice 1 routes asserted-edge mutations through the controlled
intake seam:

- EdgeStore moves from src/lithos/lcma/edges.py to a new top-level peer
  module src/lithos/edge_store.py. The lcma/edges.py file keeps the
  corpus-to-edges projection helpers and re-imports EdgeStore from the
  new location.
- CorpusIntake gains EdgeRequest / EdgeOutcome frozen dataclasses and an
  async assert_edge(agent, request) method following the same four-step
  pipeline as write / delete: ensure_agent_known (unconditional) →
  atomic EdgeStore.upsert → no view-sync step (no edge-derived view
  exists) → non-propagating EDGE_UPSERTED emission via _emit.
- LithosServer.initialize() constructs the EdgeStore directly and
  injects the same instance into ProvenanceProjection.create(...) and
  CorpusIntake(...). The self.edge_store = self.projection._edge_store
  reach-through is gone.
- CognitiveMemory declares CorpusIntake as a constructor dependency;
  edge_upsert becomes a one-call wrapper that builds an EdgeRequest and
  calls self._intake.assert_edge.
- The lithos_edge_upsert MCP handler passes provenance_actor (with a
  tool-name fallback) as the agent so ensure_agent_known runs
  unconditionally — no new MCP surface field.
- Tests cover agent auto-registration, the upsert/translate path, event
  emission ordering, and event-emit-failure non-propagation. The
  symmetric guarantee is locked in: a projection-class edge survives an
  assert_edge call at a related natural key in the same (from, to, ns)
  group.

Closes #263.
